Remove an extraneous session ping call
The yubihsm-shell appears to send an Echo(0xff) command before every
command sent. My best guess is that this is a cross-thread wakeup,
intended to poke a transport thread into action from the main UI loop.

But, as a result, all the logs recorded from the shel include these
extra ping command. Consequently this extra ping was baked into the
logic to load a key by its label. This is because this logic consists
of a sequence of shell commands, and thus the extra ping is present.

Since it isn't necessary for proper functioning, and adds unneeded
overhead, let's remove it. The code change is trivial, but all logs
with loaded keys must be regenerated with the ping command omitted.

In this case, rather than re-run all the commands manually against
an HSM, instead decrypt and replay the session messages, omit any
ping commands, and then re-encrypt the retained commands.