Make consist use of crypto.PrivateKey APIs
Implement the crypto.PrivateKey pseudo-methods Equal() and Public() on
the KeyPair type allowing its use wherever a private key is needed.
Also add the Equal() method to the cryptoSigner and cryptoDecrypter
types so these implement crypto.PrivateKey as well.

Then, add an array of tests to ensure things are implemented correctly,
and we don't accidentally e.g. pass a value receiver where a pointer
receiver is required.