Make device info trusted status a private member
Exposing the `DeviceInfo.Trusted` field was incongruous for a couple
reasons. First, this field isn't sent from the HSM and is instead
generated locally dependent upon the authentication state of the
session. All other fields are values received from the HSM

Second, exporting this as a public field allows a misbehaving
application to modify the value, and thus change the trust level of the
data.

Instead, hide this behind a private field and a public `IsTrusted`
accessor.

While in there, use Yubico's standard naming for the audit log names.